refactor: 드래그 바운더리 지정
