+++ package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@dnd-kit/core": "^6.1.0",
+    "@dnd-kit/modifiers": "^7.0.0",
     "@dnd-kit/sortable": "^8.0.0",
     "@hookform/resolvers": "^3.4.2",
     "@radix-ui/react-label": "^2.0.2",
@@ -28,8 +29,8 @@
     "react": "^18",
     "react-daum-postcode": "^3.1.3",
     "react-dom": "^18",
-    "react-multi-carousel": "^2.8.5",
     "react-hook-form": "^7.51.5",
+    "react-multi-carousel": "^2.8.5",
     "tailwindcss-animate": "^1.0.7",
     "uuidv4": "^6.2.13",
     "zod": "^3.23.8"
